Fix up replacement semantics
Don't replace on `create`/`delete` changes.  Instead, introduce a new `replaceOnChanges` property that can be used to indicate dependencies that force replacement.  (Similar to `keepers` in the `random` package or `triggers` in the Terraform `null_resource` provider).

This keeps `connection` as forcing replacement.  It is possible that should be tweaked to not always force a replacement or only force a replacement on `host` changing, but in general if the connection changes, it means the target of the command has changed, which should require re-running.

Also removes the unused `update` property from the schema.

Adds some richer testing to confirm replacement behaviour.