Improve stdout/stderr handling
This overhauls the treatment of stdout and stderr for local.Command and remote.Command.

The new rules are:
1. stdout and stderr are both rendered directly to ephemeral logs always.
2. when the command fails, the resource creation fails, and the returned error message includes the full text of stdout+stderr.
3. the stdout and stderr properties of the resource are still populated with the values of those output streams if the command succeeds.

In the future, we may want to support an option to suppress (1) for cases where the command output if overly verbose, but it is better to include it by default (similar to the docker.Image component).