Combine stdout and stderr writes
Using MultiReader overly serializes the stdout and stderr streams.  Instead, use MultiWriter to write data into multiple streams synchronously.

It appears it is not strictly possible with os.Exec to accomplish both:
1. Accessing stdout and stderr streams seperately
2. Accessing a single absolutely ordered stream for the combined stdout+stderr data

This means we can never be completely correct here - but this appears to be about as good as is possible in Go.