Add DialErrorLimit as parameter for remote command
Previously there was a hardcoded value of 10 attemps on dialing the remote host, if 10 attemps returned an error the command execution fails. Under some circustamnces it could be expected there would be a long period where the host is not responsive but the execution should keep (failing it the case based on defined timeouts)

This change just enable to adjust the attempts when getting error dialing, or even just set as unlimited so it is up to the setted up timeout to finish the execution.