Add DialErrorLimit as parameter for remote command (#180)
Previously there was a hardcoded value of 10 attempts on dialing the remote host. If all 10 attempts returned an error then command execution would fail. This change enables to adjust the number of attempts when getting an error while dialing, or even to remove the limit to keep retrying until timeout.