Regenerate SDKs
We depend on the latest version of Pulumi for codegen. We also require that running `make
build` produces zero diffs on PRs. This is great... except when Pulumi changes the output
of `make codegen`. This PR brings our checked in SDKs back into alignment with what `make
build` is generating.