Update tests
The change to both test cases adds an unknown environment property map to our expected
output. This update is an improvement in the quality of the provider, since the test gives
"environment" as an unknown input, we should see it as an unknown output as well.

We didn't in the past because we dropped computed/secret information for types that we
didn't use and that had empty values. This change represents a more accurate provider
result.