Use default implementation for input-output dependencies
WireDependencies has quite a but of subtlety at this point. It needs to account for
flowing computedness and secretness between inputs and outputs, each of which can be
handled either together or separately. When this provider was implemented, the default
implementation didn't work as expected, and didn't cover cases such as pass through
inputs.

The pulumi-go-provider now does a much better job by default, so we can rely on that
implementation. We expect the following semantics:
- If any input is computed, then all outputs are computed.
  - This does not apply to an output `k` if:
    - There is an input also called `k` and `k` is not computed.
    - `input[k] == output[k]`.
- If any input is secret, then the matching output of the same name is also secret.