Move the Logging type to work around Python codegen issue 16221 (#448)
The `Logging` enum defined in the `common` package was not correctly
exposed in the Python SDK due to pulumi/pulumi#16221, causing the error
```
AttributeError: module 'pulumi_command.common' has no attribute 'Logging'
```

As a workaround, this PR moves the Logging enum into both the `local`
and the `remote` package. This duplicates some code but keeps the
packages separated. As the provider is not GA yet, we could reverse this
once the upstream is fixed, at the cost of a breaking change.

Tested a local `pulumi up`.

Fixes https://github.com/pulumi/pulumi-command/issues/445