Correct struct tag and property name of Triggers replace annotation (#452)
The struct tag pulumi-go-provider looks for is `provider`, not `providers`.

Also, `replaceOnDelete` is not a thing. Based on the property docs
"Trigger replacements on changes to this input" I assume replace on
changes is meant. Note that two integration tests already use `Triggers`
and pass. I assume this is because a change to `Triggers` causes not a
replacement of the resource but a fresh copy?

Plus a bonus docs fix for Connection.Port.