Bring CopyFile back as deprecated resource (#473)
To allow users an easy transition to Command 1.0, especially given the
lack of alpha and beta releases, we'll keep the previous CopyFile
resource for the time being. It will be marked as deprecated, which is
currently missing in this PR, therefore it's a draft.

I also tweaked the CopyToRemote integration test for work with CopyFile
so we have coverage.