Fix repeated copying of a directory (#496)
Fixes #494

The problem was incorrect handling of repeated copies of a directory. It
would fail on the second copy because the target dir already exists,
whereas that's fine. This revealed a gap in test coverage which this PR
fills.

Also, refactor the Copy test suite so each sub-tests is independent.