Remove an unnecessary module (#619)
This module has a permanent replace pointing to `../provider`. That
effectively means this test code _should be_ part of the provider
module, but we've introduced friction around it instead.

For example this module wasn't tidied as part of
https://github.com/pulumi/pulumi-command/pull/618 because it's not
something developers expect.

The only _functional_ purpose this module serves is to hide this code
from `go test` commands in parent directories, but that can be
accomplished by a straightforward `go test .` instead of `go test
./...`.

However as far as I can tell **these tests were never actually being
run!**