Don't force updates when upgrading to 1.x (#924)
This fixes #912 by no longer forcing resources to update when upgrading
the provider.

The "default true" behavior of the `addPreviousOutputInEnv` property is
preserved, but we stop setting it by default on resources. Only when a
user explicitly provides a value will it be reflected in state.

(In hindsight this property should have been named something like
`omitPreviousOutputInEnv` to avoid this polarity issue.)

The upgrade test was modified to reproduce the problem by expecting no
changes instead of allowing some updates. The existing test wasn't
testing against the local SDK, so it was also moved to examples (which
install the SDK) and modified to actually link against it.

Fixes #912.